Extract builder methods from `PartialHalfEdge`
