+++ crates/fj-kernel/src/algorithms/sweep/face.rs
@@ -84,6 +84,7 @@ mod tests {
 
     use crate::{
         algorithms::{reverse::Reverse, transform::TransformObject},
+        builder::HalfEdgeBuilder,
         objects::{Face, HalfEdge, Objects, Sketch},
         partial::HasPartial,
     };
