+++ crates/fj-kernel/src/validate/edge.rs
@@ -196,7 +196,7 @@ mod tests {
     use fj_interop::ext::ArrayExt;
 
     use crate::{
-        builder::VertexBuilder,
+        builder::{HalfEdgeBuilder, VertexBuilder},
         objects::{GlobalCurve, HalfEdge, Objects},
         partial::HasPartial,
         validate::Validate2,
