+++ crates/fj-operations/src/sketch.rs
@@ -2,6 +2,7 @@ use std::ops::Deref;
 
 use fj_interop::{debug::DebugInfo, mesh::Color};
 use fj_kernel::{
+    builder::HalfEdgeBuilder,
     objects::{Cycle, Face, HalfEdge, Objects, Sketch},
     partial::HasPartial,
     validate::{Validate, Validated, ValidationConfig, ValidationError},
